backend/requirements.txt
blinker==1.9.0
cffi==2.0.0
click==8.1.8
cryptography==46.0.3
Flask==3.1.2
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pycparser==2.23
PyMuPDF==1.26.5
PyMySQL==1.1.2
python-dotenv==1.2.1
SQLAlchemy==2.0.45
typing_extensions==4.15.0
Werkzeug==3.1.5
zipp==3.23.0
